feat: Informative error handling
Since knowing where errors are coming from is quite hard this commit makes sure to try and report whether it's from Kubernetes or an error in the test itself.

A few improvements to the K8s and testing code in general to add more tests for errors and more granular error reporting.